Extract venv fixtures from test_distutils_adoption
… and change it to install the pre-build setuptools wheel (fixture)
instead of installing from the source tree
