+++ setuptools/tests/test_distutils_adoption.py
@@ -1,37 +1,13 @@
 import os
 import sys
 import functools
-import subprocess
 import platform
 import textwrap
 
 import pytest
-import jaraco.envs
-import path
 
 
 IS_PYPY = '__pypy__' in sys.builtin_module_names
-
-
-class VirtualEnv(jaraco.envs.VirtualEnv):
-    name = '.env'
-    # Some version of PyPy will import distutils on startup, implicitly
-    # importing setuptools, and thus leading to BackendInvalid errors
-    # when upgrading Setuptools. Bypass this behavior by avoiding the
-    # early availability and need to upgrade.
-    create_opts = ['--no-setuptools']
-
-    def run(self, cmd, *args, **kwargs):
-        cmd = [self.exe(cmd[0])] + cmd[1:]
-        return subprocess.check_output(cmd, *args, cwd=self.root, **kwargs)
-
-
-@pytest.fixture
-def venv(tmp_path, tmp_src):
-    env = VirtualEnv()
-    env.root = path.Path(tmp_path / 'venv')
-    env.req = str(tmp_src)
-    return env.create()
 
 
 def popen_text(call):
